Lib

Description

Documentation

type Key = Char

type Occurences = Map Key Integer

type Letter = Char

type Alphabet = Map Letter Occurences

occurences :: Occurences

initOccurences :: Occurences

Initializes a new (inner) list of Char, Int tuples.

initAlphabet :: Alphabet

Initializes a new (outer) list of Char, Occurences tuples.

frequency :: Ord a => [a] -> [(a, Int)]

Used to get the number of all occurences of an datatype.

lookup :: Char -> [(Char, Int)] -> Int

Used by getCharCountValue to get the number of occurences from a specific character.

getCharCountValue :: Char -> [(Char, Int)] -> Int

Returns the number of occurences from a specific character.

swapTupleValues :: ((Char, Char), Int) -> (Char, (Char, Int))

Swaps the three values of a tuple of tuples.

changeIntTupleToFloat :: [(Char, Int)] -> (Char, (Char, Int)) -> (Char, (Char, Float))

Changes the old list with int occurences to a new list with float percentages (0.0 - 1.0).

countLetters :: String -> Char -> Int

Counts the occurences of a char value in a string.

printFloatTuple :: (Char, (Char, Float)) -> String

Prints a tuple after the occurence count has been changed to the percentages (0.0 - 1.0).

groupTupleLists :: [(Char, (Char, Float))] -> [[(Char, (Char, Float))]]

Changes the list's structure.

kummulate :: [(Char, Float)] -> [(Char, Float)]

Each character has it's own area between 0.0 and 1.0. This is used to ensure better result from our randomizer.

kummulateAddNumber :: (Char, Float) -> Float -> (Char, Float)

Returns the sum of both tuple values and another value.

getKeyOfTuple :: [(Char, (Char, Float))] -> Char

Gets the key of a tuple.

getSubTupelOfTuple :: (Char, (Char, Float)) -> (Char, Float)

Gets the inner tuple of a tuple.

makeListOfTuple :: [(Char, (Char, Float))] -> [(Char, Float)]

Changes the list to a list of tuples.

makeCharTupleListTuple :: [(Char, (Char, Float))] -> (Char, [(Char, Float)])

Changes the list of tuples a tuples with a list as the second argument.

openFile :: String -> IO ()

Used to start the whole process.

charToString :: Char -> String

Changes a char to a string.

getRandomNumberOneTo :: IO Int

Returns a random Int (1-1000).

optimizeList :: [[(Char, (Char, Float))]] -> [(Char, [(Char, Float)])]

Changes the list's structure for further use.

generateText :: String -> Int -> Char -> [(Char, [(Char, Float)])] -> IO ()

Generates a text based on random numbers and the occurence values of the characters.